split per-method handlers into separate files, dynamically loaded. - fixed web3 ipc socket response handlng - finalized sendRequest() initial draft
